feat: Split InputManifest::load* into attached/detached forms
Previously, the `load`, `load_async` and `sha256` constructors for
`InputManifest` all took a second parameter that was an `Option` that could
be a value used to identify the target artifact, such as a `&Path`. In practice,
this lead to awkward type inference errors, requiring callers to do things like
`None::<&String>`, which we don't really want them to have to do.

These 3 constructors have now been split into 3 pairs of constructors, a regular
one which takes a non-optional target (which, because it's not optional, means
inference should always have enough information to give useful errors), and
another which does not take a target and is explicitly for constructing a
"detached" manifest.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>