feat: Simplified `Storage` trait
This introduces a new `Match` type enabling both the `get_*` and `remove_*`
methods to be reduced to single `get_manifest` and `remove_manifest` methods.

`Match` controls how the matching is done for those operations, either by
matching the target's Artifact ID, or by matching the Artifact ID of the
manifest itself.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>